requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.3
opencv-python==4.8.1.78
pillow==10.0.0
numpy==1.26.4
scipy==1.11.0
scikit-image==0.21.0
pytest==7.4.0
pytest-benchmark==4.0.0
fastapi==0.103.0
uvicorn==0.23.0
python-multipart==0.0.6
pydantic==2.3.0
PyMuPDF==1.23.0
pymysql==1.1.0
